new: input dto 에 보너스 볼 추가
